Moved fakeProperty into the main package
